database creation and manipulation in api instead of app
